Fixed compiler warnings

+++ src/FrameData.h
@@ -11,8 +11,8 @@ namespace OrientView
 	struct FrameData
 	{
 		uint8_t* data = nullptr;	// Raw data, format depends on context (RGBA32 or GRAY8)
-		int dataLength = 0;			// Data length in bytes
-		int rowLength = 0;			// Length of the row in bytes (could be larger than width)
+		size_t dataLength = 0;		// Data length in bytes
+		size_t rowLength = 0;		// Length of the row in bytes (could be larger than width)
 		int width = 0;				// Width in pixels
 		int height = 0;				// Height in pixels
 		int duration = 0;			// Duration in microseconds
