+++ src/Mp4File.h
@@ -18,7 +18,7 @@ namespace OrientView
 		bool open(const QString& fileName);
 		bool setParam(x264_param_t* param);
 		bool writeHeaders(x264_nal_t* nal);
-		bool writeFrame(uint8_t* payload, int size, x264_picture_t* picture);
+		bool writeFrame(uint8_t* payload, size_t size, x264_picture_t* picture);
 		bool close(int64_t lastPts);
 		
 	private:
